refactor: introduce TopicPool
